+++ gallerygather/gallerygather/src/main/java/com/kosa/gallerygather/entity/Exhibition.java
@@ -63,6 +63,6 @@ public class Exhibition {
 
     @Column(name = "avg_rating")
     @ColumnDefault("0")
-    private Integer avgRating;
+    private Double avgRating;
 
 }
